chore: extract provider code to a separate file
